fix: 修改baseline文件修改逻辑
只有系统更新时，修改baseline文件

Log: 修改baseline文件修改逻辑
Task: https://pms.uniontech.com/task-view-303345.html
Influence:

+++ src/lastore-daemon/message_report.go
@@ -222,11 +222,14 @@ func (m *UpdatePlatformManager) UpdateBaseline() {
 
 // 进行安装更新前，需要复制文件替换软连接
 func (m *UpdatePlatformManager) replaceVersionCache() {
-	err := os.RemoveAll(cacheVersion)
-	if err != nil {
-		logger.Warning(err)
+	if utils.IsSymlink(cacheVersion) {
+		// 如果cacheVersion已经是文件了，那么不再用源文件替换
+		err := os.RemoveAll(cacheVersion)
+		if err != nil {
+			logger.Warning(err)
+		}
+		copyFile(realVersion, cacheVersion)
 	}
-	copyFile(realVersion, cacheVersion)
 }
 
 // 安装更新并检查完成后，需要用软连接替换文件
